Fixed bug and changed example page
