--- examples/next/app/next13app/bug/page.tsx
@@ -1,5 +1,0 @@
-import BugExample from './BugExample';
-
-export default function Home() {
-  return <BugExample />;
-}
