+++ examples/next/app/next13app/call-api-route/page.tsx
@@ -1,0 +1,5 @@
+import ApiRouteExample from './ApiRouteExample';
+
+export default function Home() {
+  return <ApiRouteExample />;
+}
